+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - hours-within-year.rb [todo from chapter-1]
 - minutes-within-decade.rb [todo from chapter-1]
 - seconds-old.rb [todo from chapter-1]
+- chocolates-within-life.rb [todo from chapter-1]